+++ src/cli/schemas/index.ts
@@ -2,7 +2,13 @@ export * from './base';
 export * from './checkConfig';
 export * from './generate';
 export * from './init';
+export * from './previewChanges';
 export * from './updateConfig';
 
 // Тип для всех возможных CLI опций (для union type)
-export type CLIOptions = import('./generate').GenerateOptions | import('./checkConfig').CheckConfigOptions | import('./updateConfig').UpdateConfigOptions | import('./init').InitOptions;
+export type CLIOptions =
+    | import('./generate').GenerateOptions
+    | import('./checkConfig').CheckConfigOptions
+    | import('./updateConfig').UpdateConfigOptions
+    | import('./init').InitOptions
+    | import('./previewChanges').TPreviewChangesOptions;