+++ src/cli/schemas/init.ts
@@ -1,13 +1,12 @@
 import { z } from 'zod';
 
-import { DEFAULT_CUSTOM_REQUEST_PATH, DEFAULT_SPECS_DIR } from '../../common/Consts';
-import { baseCLIOptionsSchema } from './base';
+import { baseCLIOptionsSchema, emptyStringToUndefined } from './base';
 
 export const initOptionsSchema = baseCLIOptionsSchema.extend({
-    specsDir: z.string().min(1).optional().default(DEFAULT_SPECS_DIR),
-    request: z.string().min(1).optional().default(DEFAULT_CUSTOM_REQUEST_PATH),
-    useCancelableRequest: z.boolean().optional().default(false),
-    useInteractiveMode: z.boolean().optional().default(false),
+    specsDir: emptyStringToUndefined,
+    request: emptyStringToUndefined,
+    useCancelableRequest: z.boolean().optional(),
+    useInteractiveMode: z.boolean().optional(),
 });
 
 export type InitOptions = z.infer<typeof initOptionsSchema>;