+++ src/common/VersionedSchema/MultiOptionsVersioned/MultiOptionsSchemaV5.ts
@@ -1,28 +1,38 @@
-import Joi from 'joi';
+import {z} from 'zod';
 
-import { HttpClient } from '../../../core';
-import { additionalParametersSchemaV2, experimentalParametersSchemaV2, outputPathsSchema, specialParametersSchemasV2 } from '../CommonSchemas';
-import { mergeObjectSchemas } from '../Utils/mergeObjectSchemas';
+import { additionalParametersSchemaV2, experimentalParametersSchemaV2, specialParametersSchemasV2 } from '../CommonSchemas';
+import { multiOptionsSchemaV4 } from './MultiOptionsSchemaV4';
 
-/**
- * The scheme of a set of generator options (v5)
- */
-export const multiOptionsSchemaV5 = mergeObjectSchemas(
-    Joi.object({
-        httpClient: Joi.string().valid(...Object.values(HttpClient)),
-        items: Joi.array()
-            .items(
-                mergeObjectSchemas(
-                    Joi.object({
-                        input: Joi.string().required(),
-                    }),
-                    outputPathsSchema
-                )
-            )
-            .min(1)
-            .required(),
-    }),
-    specialParametersSchemasV2,
-    additionalParametersSchemaV2,
-    experimentalParametersSchemaV2
-);
+const multiOptionsBaseV5 = multiOptionsSchemaV4.pick({
+    items: true,
+    httpClient: true,
+    request: true,
+    interfacePrefix: true,
+    enumPrefix: true,
+    typePrefix: true,
+});
+
+export const multiOptionsSchemaV5 = z.object({
+    ...multiOptionsBaseV5.shape,
+    ...specialParametersSchemasV2.shape,
+    ...additionalParametersSchemaV2.shape,
+    ...experimentalParametersSchemaV2.shape,
+});
+
+/*
+type TMulti = {
+    items: { ... 6 more }[];
+    useCancelableRequest: boolean | undefined;
+    sortByRequired: boolean | undefined;
+    useSeparatedIndexes: boolean | undefined;
+    request: string | undefined;
+    interfacePrefix: string | undefined;
+    enumPrefix: string | undefined;
+    typePrefix: string | undefined;
+    useOptions: boolean | undefined;
+    useUnionTypes: boolean | undefined;
+    excludeCoreServiceFiles: boolean | undefined;
+    includeSchemasFiles: boolean | undefined;
+    httpClient: HttpClient.FETCH | HttpClient.XHR | HttpClient.NODE | HttpClient.AXIOS | undefined;
+}
+*/