+++ src/common/VersionedSchema/OptionsVersioned/OptionsSchemaV4.ts
@@ -1,19 +1,36 @@
-import Joi from "joi";
+import { z } from "zod";
 
 import { HttpClient } from '../../../core';
 import { additionalParametersSchemaV2, experimentalParametersSchemaV2, outputPathsSchema, specialParametersSchemasV2 } from '../CommonSchemas';
-import { mergeObjectSchemas } from '../Utils/mergeObjectSchemas';
 
-/**
- * The scheme of a set of generator options (Версия 4).
- */
-export const optionsSchemaV4 = mergeObjectSchemas(
-    Joi.object({
-        input: Joi.string().required(),
-        httpClient: Joi.string().valid(...Object.values(HttpClient)),
-    }),
-    outputPathsSchema,
-    specialParametersSchemasV2,
-    additionalParametersSchemaV2,
-    experimentalParametersSchemaV2,
-);
+export const optionsSchemaV4 = z.object({
+    ...outputPathsSchema.shape,
+    ...specialParametersSchemasV2.shape,
+    ...additionalParametersSchemaV2.shape,
+    ...experimentalParametersSchemaV2.shape,
+    input: z.string().min(1),
+    httpClient: z.enum(HttpClient),
+});
+
+/*
+type TOptions = {
+    input: string;
+    httpClient: HttpClient.FETCH | HttpClient.XHR | HttpClient.NODE | HttpClient.AXIOS;
+    output: string;
+    useCancelableRequest: boolean | undefined;
+    sortByRequired: boolean | undefined;
+    useSeparatedIndexes: boolean | undefined;
+    request: string | undefined;
+    interfacePrefix: string | undefined;
+    enumPrefix: string | undefined;
+    typePrefix: string | undefined;
+    useOptions: boolean | undefined;
+    useUnionTypes: boolean | undefined;
+    excludeCoreServiceFiles: boolean | undefined;
+    includeSchemasFiles: boolean | undefined;
+    outputCore: string | undefined;
+    outputServices: string | undefined;
+    outputModels: string | undefined;
+    outputSchemas: string | undefined;
+}
+*/