+++ src/common/VersionedSchema/Types.ts
@@ -1,11 +1,12 @@
-import Joi from 'joi';
+import { z } from 'zod';
 
 import { EVersionedSchemaType } from './Enums';
 
-export type VersionedSchema<T> = {
-    schema: Joi.Schema<T>;
-    type: EVersionedSchemaType;
+export type VersionedSchema<TSchema extends z.ZodTypeAny> = {
+    schema: TSchema;
+    baseSchema: z.ZodObject<any>;
     version: string;
+    type: EVersionedSchemaType;
 };
 
 export type SchemaMigrationPlan<From, To> = {
@@ -18,9 +19,6 @@ export type SchemaMigrationPlan<From, To> = {
      */
     description?: string;
 };
-
-// A universal function for traversing the structure of an object with result typing
-export type TraverseHandler<T> = (value: any, recurse: (v: any) => void, result: T) => boolean;
 
 export type VersionMatchResult = {
     lastVersionIndex: number;