+++ src/core/api/v2/parser/getOperationParameter.ts
@@ -5,7 +5,6 @@ import { getEnum } from '../../../utils/getEnum';
 import { getEnumFromDescription } from '../../../utils/getEnumFromDescription';
 import { getOperationParameterName } from '../../../utils/getOperationParameterName';
 import { getPattern } from '../../../utils/getPattern';
-import { normalizeRef } from '../../../utils/normalizeRef';
 import { Parser } from '../Parser';
 import type { OpenApi } from '../types/OpenApi.model';
 import type { OpenApiParameter } from '../types/OpenApiParameter.model';
@@ -48,8 +47,7 @@ export function getOperationParameter(this: Parser, openApi: OpenApi, parameter:
     };
 
     if (parameter.$ref) {
-        const normalizedRef = normalizeRef(parameter.$ref, parentRef);
-        const definitionRef = this.getType(parameter.$ref, normalizedRef);
+        const definitionRef = this.getType(parameter.$ref, parentRef);
         operationParameter.export = 'reference';
         operationParameter.type = definitionRef.type;
         operationParameter.base = definitionRef.base;
@@ -109,8 +107,7 @@ export function getOperationParameter(this: Parser, openApi: OpenApi, parameter:
 
     if (parameter.schema) {
         if (parameter.schema.$ref) {
-            const normalizedRef = normalizeRef(parameter.schema.$ref, parentRef);
-            const model = this.getType(parameter.schema.$ref, normalizedRef);
+            const model = this.getType(parameter.schema.$ref, parentRef);
             operationParameter.export = 'reference';
             operationParameter.type = model.type;
             operationParameter.base = model.base;