+++ src/core/api/v2/parser/getOperationResponse.ts
@@ -2,7 +2,6 @@ import { safeHasOwn } from '../../../../common/utils/safeHasOwn';
 import type { OperationResponse } from '../../../types/shared/OperationResponse.model';
 import { getComment } from '../../../utils/getComment';
 import { getPattern } from '../../../utils/getPattern';
-import { normalizeRef } from '../../../utils/normalizeRef';
 import { Parser } from '../Parser';
 import type { OpenApi } from '../types/OpenApi.model';
 import type { OpenApiResponse } from '../types/OpenApiResponse.model';
@@ -36,8 +35,7 @@ export function getOperationResponse(this: Parser, openApi: OpenApi, response: O
     // then we need to parse the schema!
     if (response.schema) {
         if (response.schema.$ref) {
-            const normalizedRef = normalizeRef(response.schema.$ref, parentRef);
-            const model = this.getType(response.schema.$ref, normalizedRef);
+            const model = this.getType(response.schema.$ref, parentRef);
             operationResponse.export = 'reference';
             operationResponse.type = model.type;
             operationResponse.base = model.base;